Update request.js documentation
This clarifies how to send form data. `.field()` works for multi-part forms, but according to the [SuperAgent documentation](https://visionmedia.github.io/superagent/#post-/-put-requests) `.send()` should be used to send the post data as `application/x-www-form-urlencoded`. This addresses issue #86. There is an older pull request by @BenAychh, but it modified the README.md file directly. According to the discussion, the README.md file is generated based on request.js, and so the documentation should be modified here.