fix: add superagent types as prod dependency
We expose superagent's types in our own types, so this should be a
production dependency. Otherwise, we will cause errors in the typescript
builds of consumers (unless they skip lib check).

It does mean we add an unnecessary dependency for non-ts people, but
there's not any simple way around that.