test: rework tests to node-only
Long ago it seems we stopped testing in a browser, so this removes the
associated switches in the test files and moves to a node-only suite.

Meanwhile, this moves to using c8 for coverage and also reworks the
tests to hit a local server rather than a real one (previously was
hitting `chaijs.com` and another domain or two which changed underneath
it over time).